+++ manual/src/changelog.md
@@ -2,8 +2,12 @@
 
 ## 0.2.0 (unreleased)
 
+- Validate service URLs before save using URL parse, improved URL error handling.
+- Fix inconsistent core error display strings.
+- Fix duplicate user email address returned internal server error code.
+
 ---
 
 ## 0.1.0 (2019-05-16)
 
-First version.
+- First version.